+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+------------
+### Added
+- Added a `readme-md init` command to initialize a project with a default config.
+
 [0.8.1] - 2018-01-13
 --------------------
 ### Fixed
@@ -76,6 +81,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
+[Unreleased]: https://github.com/jbenner-radham/node-readme-md-cli/compare/v0.8.1...HEAD
 [0.8.1]: https://github.com/jbenner-radham/node-readme-md-cli/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/jbenner-radham/node-readme-md-cli/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/jbenner-radham/node-readme-md-cli/compare/v0.6.1...v0.7.0